Adjust the project structure, and allow installation via pip
