+++ .gitignore
@@ -3,7 +3,7 @@ conv.txt
 test*
 
 # Byte-compiled / optimized / DLL files
-__pycache__/
+*__pycache__/
 *.py[cod]
 *$py.class
