+++ gpttrace/__init__.py
@@ -1,0 +1,2 @@
+from gpttrace.GPTtrace import main
+__version__ = "0.1.2"
