+++ gpttrace/__main__.py
@@ -1,0 +1,4 @@
+# This file allows the construction of executable file. If there is no such file, only the package is generated.
+from gpttrace.GPTtrace import main
+
+main()
