+++ gpttrace/cmd.py
@@ -1,13 +1,18 @@
 import subprocess
 import openai
 import json
-from gen_func_call import gen_func_call
 
-def cmd_parser(cmd: str, query: str, verbose=False) -> None:
+from langchain.chains.conversation.memory import ConversationBufferMemory
+from langchain.chat_models import ChatOpenAI
+from langchain.chains import ConversationChain
+from gpttrace.utils.prompt import func_call_prompt
+from gpttrace.config import cfg
+
+def cmd(cmd_name: str, query: str, verbose=False) -> None:
     """
     Generate the command based on query and execute the command.
 
-    :param cmd: name of command.
+    :param cmd_name: name of command.
     :param query: The task that the user wants to accomplish with `cmd`.
     :param verbose: Whether to print extra information.
     """
@@ -15,15 +20,15 @@ def cmd_parser(cmd: str, query: str, verbose=False) -> None:
     functions = get_predifine_funcs()
     for func in functions:
         # Gets the predefined function call
-        if func["name"] == cmd:
-            func_call= func
+        if func["name"] == cmd_name:
+            func_call = func
             break
     if func_call is None:
         # Generate a function call based on a given command
-        func_call = gen_func_call(cmd, verbose)
+        func_call = gen_func_call(cmd_name, verbose)
     messages = [{"role": "user", "content": query}]
     response = openai.ChatCompletion.create(
-        model="gpt-3.5-turbo",
+        model=cfg.get("DEFAULT_MODEL"),
         messages=messages,
         functions=[func_call],
         function_call="auto",
@@ -36,6 +41,7 @@ def cmd_parser(cmd: str, query: str, verbose=False) -> None:
         exec_cmd(cmd_name, args, func_descript)
     else:
         print("LLM does not call any bcc tools.")
+
 
 def exec_cmd(cmd_name: str, args: str, func_descript: json) -> None:
     """
@@ -67,43 +73,75 @@ def exec_cmd(cmd_name: str, args: str, func_descript: json) -> None:
     print("\u001b[1;32m", "Run: ", " ".join(full_command), "\u001b[0m")
     try:
         subprocess.run(full_command, text=True, check=True)
-    except Exception as e:
-        print("\u001b[1;32m\bFailed to execute command!\u001b[0m")
-        print(e)
+    except subprocess.CalledProcessError as err:
+        print("\u001b[1;31m\bFailed to execute command!\u001b[0m")
+        print(err)
 
 def get_predifine_funcs() -> str:
     """
     Gets the JSON format description information of the predefined function call.
-    
+
     :return: List of JSON-formatted function calls.
     """
-    with open('./funcs.json', 'r') as file:
+    func_path = cfg.get("BCC_FUNC_CALL_PATH")
+    with open(func_path, 'r', encoding='utf-8') as file:
         data = file.read()
     functions = json.loads(data)
     return functions
 
-def get_specify_func(functions, cmd) -> json:
+def gen_func_call(cmd_name: str, verbose: bool) -> str:
+    """
+    Generates a funciton call for the given command.
+
+    :param cmd_name: Name of command.
+    :verbose: Whether to print extra information.
+    :return: The function call in JSON format corresponding to the command.
+    """
+    model_name = cfg.get("DEFAULT_MODEL")
+    llm = ChatOpenAI(model=model_name,temperature=0)
+    agent_chain = ConversationChain(llm=llm, verbose=verbose,
+                    memory=ConversationBufferMemory())
+    help_doc = get_command_help(cmd_name)
+    prompt = func_call_prompt(cmd_name, help_doc)
+    response = agent_chain.predict(input=prompt)
+    return response
+
+def get_command_help(command) -> str:
+    """
+    Gets help documentation for the command.
+
+    :param command: Name of command.
+    :return: Help documentation for the command.
+    """
+    try:
+        output = subprocess.check_output([command, '--help'], universal_newlines=True)
+        return output
+    except subprocess.CalledProcessError as err:
+        return f"Error executing help command: {err.output}"
+
+def get_specify_func(functions, cmd_name) -> json:
     """
     Gets the JSON format description of the function call for the specified command.
-    
+
     :param functions: A predefined list of function calls.
-    :param cmd: A specific command.
+    :param cmd_name: A specific command.
     :return func: The function call corresponding to cmd.
     """
     for func in functions:
-        if func.get('name') == cmd:
+        if func.get('name') == cmd_name:
             return func
     return None
 
-def is_positional_arg(cmd, arg) -> bool:
-    """
-    Determine whether the argument arg in the command cmd is a positional argument.
 
-    :param cmd: name of command.
+def is_positional_arg(cmd_name, arg) -> bool:
+    """
+    Determine if the argument passed to the command is a position argument.
+
+    :param cmd_name: name of command.
     :param arg: argument passed to command.
     :return: Whether arg is a positional parameter of cmd.
     """
-    
+
     positional_dict = {
         "biolatency-bpfcc": ["interval", "count"],
         "biotop-bpfcc": ["interval", "count"],
@@ -185,9 +223,10 @@ def is_positional_arg(cmd, arg) -> bool:
         "zfsslower-bpfcc": ["min_ms"],
     }
 
-    if positional_dict.get(cmd) is not None:
+    if positional_dict.get(cmd_name) is not None:
         return arg in positional_dict[cmd]
     return False
 
+
 if __name__ == "__main__":
-    cmd_parser("print 1 second summaries, 10 times", True)
+    cmd("print 1 second summaries, 10 times", True)
