+++ gpttrace/utils/__init__.py
@@ -1,0 +1,5 @@
+import os
+import sys
+
+project_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+sys.path.append(project_root)
